+++ services/workflows-service/src/workflow/workflow.controller.internal.intg.test.ts
@@ -33,6 +33,7 @@ import { Request } from 'express';
 import { WorkflowDefinitionRepository } from '@/workflow-defintion/workflow-definition.repository';
 import { UiDefinitionService } from '@/ui-definition/ui-definition.service';
 import { UiDefinitionRepository } from '@/ui-definition/ui-definition.repository';
+import { BusinessService } from '@/business/business.service';
 
 describe('/api/v1/internal/workflows #api #integration', () => {
   let app: INestApplication;
@@ -60,6 +61,7 @@ describe('/api/v1/internal/workflows #api #integration', () => {
       StorageService,
       WorkflowEventEmitterService,
       BusinessRepository,
+      BusinessService,
       WorkflowDefinitionRepository,
       WorkflowRuntimeDataRepository,
       WorkflowService,